+++ packages/common/src/index.ts
@@ -1,2 +1,3 @@
+export * from "./requirements";
 export * from "./requirements-txt";
 export * from "./promise-delegate";