+++ packages/mountable/config/webpack.config.js
@@ -306,8 +306,6 @@ module.exports = function (webpackEnv) {
         // Support React Native Web
         // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
         'react-native': 'react-native-web',
-        // Stlite: To resolve the alias streamlit/frontend uses
-        "src": path.resolve(__dirname, "../../../streamlit/frontend/src"),
         // Allows for better profiling with ReactDevTools
         ...(isEnvProductionProfile && {
           'react-dom$': 'react-dom/profiling',