+++ packages/sharing-editor/public/samples/011_component_gallery/pages/utilities.html.py
@@ -1,0 +1,11 @@
+import streamlit as st
+
+code = """
+<style>
+    p {
+        color: red;
+    }
+</style>
+"""
+st.html(code)
+st.markdown("Lorem ipsum")